Fix bugs and update the dependency
Fix bug that do not update the database after creating a transaction that  must transfer some coins to other, but do no insert the serial number to utxo ring entry, and not move thr unspent txo entries to spent but unmined bucket.
Fix a bug that incurs panic when copy a utxoring to update the block input bucket entries.
Update the dependency.