Refracting
(1) remove some outdate comments
(2) Add a const variable for seed length
(3) Disable some method unmatched abelian
(4) Remove some useless bucket in database
(5) Remove payee manager which acts as a manager to manage the derivation of master address when enabled salrs. But now it switch to pqringct.
(6) Remove the goroutine ring fresh which acts as an updater for updating the serial number when enabled salrs. But now it switch to pqringct.
(7) Add some comment and todo list